dataset/requirements.txt
# <Socketio Client>
python-engineio==4.12.2
python-socketio==5.13.0
requests==2.32.4
# </Socketio Client>

# <Miscellenaous>
dotenv==0.9.9
pandas==2.3.2
#httpx==0.28.1
# </Miscellenaous>

